requirements.txt
requests
newsapi-python
openai
python-dotenv
cloudinary
feedparser
python-dateutil
